src/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
sqlalchemy>=2.0.0
pyinstaller>=6.0.0
